@@ -1,50 +1,146 @@
-[tool.poetry]
+[project]
 name = "kegstandcli"
 version = "0.3.9"
 description = "The Developer's Toolbelt For Accelerating Mean-Time-To-Party on AWS"
-authors = ["JensRoland <mail@jensroland.com>"]
-license = "MIT"
-repository = "https://github.com/jensroland/kegstand"
-homepage = "https://kegstand.dev"
+authors = [
+    { name = "JensRoland", email = "mail@jensroland.com" }
+]
+license = { text = "MIT" }
 readme = "README.md"
-packages = [
-    { include = "kegstandcli", from = "src" },
-]
-# Keywords description https://python-poetry.org/docs/pyproject/#keywords
-keywords = []  #! Update me
-
-# Pypi classifiers: https://pypi.org/classifiers/
+requires-python = ">=3.12"
+keywords = []
 classifiers = [
-  "Development Status :: 4 - Beta",
-  "Intended Audience :: Developers",
-  "Operating System :: OS Independent",
-  "Topic :: Software Development :: Libraries :: Python Modules",
-  "License :: OSI Approved :: MIT License",
-  "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "Operating System :: OS Independent",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.12",
+]
+dependencies = [
+    "aws-cdk-lib>=2.67.0",
+    "aws-solutions-constructs-aws-apigateway-lambda>=2.39.0",
+    "boto3>=1.17.113",
+    "click>=8.0.3",
+    "constructs>=10.0.0,<11.0.0",
+    "copier>=9.4.1",
+    "directory-tree>=1.0.0",
+    "pyjwt>=2.1.0",
+    "tomlkit>=0.11.7",
+    "xxhash>=3.2.0",
 ]
 
-[tool.poetry.scripts]
+[dependency-groups]
+dev = [
+    "boto3-stubs>=1.35.82",
+    "mypy>=1.13.0",
+    "pytest>=7.3.0",
+    "pytest-cov>=4.0.0",
+    "pytest-subprocess>=1.5.0",
+    "ruff>=0.1.9",
+]
+
+[project.scripts]
 keg = "kegstandcli.cli.__main__:kegstandcli"
 
-[tool.poetry.dependencies]
-aws-cdk-lib = "^2.67.0"
-aws-solutions-constructs-aws-apigateway-lambda = "^2.39.0"
-boto3 = "^1.17.113"
-click = "^8.0.3"
-constructs = ">=10.0.0,<11.0.0"
-copier = "^6.2.0"  # 7.0.1 has a bug that deletes the template folder, see https://github.com/copier-org/copier/issues/1029
-pyjwt = "^2.1.0"
-python = "^3.9"
-tomlkit = "^0.11.7"
-xxhash = "^3.2.0"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^7.3.0"
-pytest-subprocess = "^1.5.0"
-sentient-switchblade = "^0.3.5"
+[project.urls]
+Homepage = "https://kegstand.dev"
+Repository = "https://github.com/jensroland/kegstand"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.pytest.ini_options]
+# Directories that are not visited by pytest collector:
+norecursedirs =["dist", "docs", ".tox", ".git", "__pycache__", ".ruff_cache"]
+pythonpath = "src/kegstandcli"
+
+[tool.coverage.report]
+fail_under = 70
+show_missing = true
+skip_empty = true
+
+[tool.ruff]
+# Exclude a variety of commonly ignored directories.
+exclude = [
+    ".bzr",
+    ".direnv",
+    ".eggs",
+    ".git",
+    ".git-rewrite",
+    ".hg",
+    ".mypy_cache",
+    ".nox",
+    ".pants.d",
+    '.pytest_cache',
+    ".pytype",
+    ".ruff_cache",
+    ".svn",
+    ".tox",
+    ".venv",
+    "__pypackages__",
+    "_build",
+    "buck-out",
+    "build",
+    "dist",
+    "node_modules",
+    "venv",
+]
+
+# Slightly longer lines are allowed because it's not 1995 anymore.
+line-length = 100
+
+# Use latest supported Python version
+target-version = "py312"
+
+[tool.ruff.lint]
+# Enable pycodestyle (`E`), Pyflakes (`F`), isort (`I`), pep8-naming (`N`), 
+# flake8-bugbear (`B`), flake8-builtins (`A`), flake8-comprehensions (`C4`),
+# flake8-pie (`PIE`), flake8-simplify (`SIM`), flake8-unused-arguments (`ARG`), 
+# pylint (`PL`), tidy-imports (`TID`), flake8-quotes (`Q`), bandit (`S`)
+select = ["E", "F", "I", "N", "B", "A", "C4", "PIE", "SIM", "ARG", "PL", "TID", "Q", "S"]
+ignore = ["ISC"]
+
+# Allow autofix for all enabled rules (when `--fix`) is provided.
+fixable = ["ALL"]
+unfixable = []
+
+# Allow unused variables when underscore-prefixed.
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+[tool.ruff.lint.mccabe]
+# Unlike Flake8, default to a complexity level of 10.
+max-complexity = 10
+
+[tool.ruff.lint.isort]
+known-first-party = ["kegstandcli"]
+
+[tool.ruff.lint.flake8-quotes]
+docstring-quotes = "double"
+inline-quotes = "double"
+
+[tool.ruff.lint.flake8-bugbear]
+extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Body"]
+
+[tool.ruff.lint.per-file-ignores]
+# Tests can use assertions and magic values
+"tests/**/*" = ["S101", "PLR2004"]
+# Templates can use magic values
+"**/*.tmpl" = ["PLR2004"]
+
+[tool.ruff.format]
+# Use double quotes for strings.
+quote-style = "double"
+# Indent with spaces, rather than tabs.
+indent-style = "space"
+# Respect magic trailing commas.
+skip-magic-trailing-comma = false
+# Automatically detect the appropriate line ending.
+line-ending = "auto"
+
+[tool.mypy]
+exclude = [
+    "^tests/test_data/",
+]
