@@ -1,10 +1,12 @@
 #!/usr/bin/env python3
-import boto3
-import click
-import os
+"""CDK app for Kegstand infrastructure."""
+
 import sys
+from pathlib import Path
 
 import aws_cdk as cdk
+import boto3
+import click
 from aws_cdk import aws_apigateway as apigw
 from aws_cdk import aws_cognito as cognito
 
@@ -30,10 +32,7 @@ parent_stack = cdk.Stack(
     app,
     parent_stack_name,
     description=f"{config['project']['name']} v.{config['project']['version']}",
-    env=cdk.Environment(
-        account=os.environ["CDK_DEFAULT_ACCOUNT"],
-        region=region
-    )
+    env=cdk.Environment(account=cdk.Aws.ACCOUNT_ID, region=region),
 )
 
 # Create stacks/modules
@@ -42,9 +41,9 @@ modules = {}
 if "cdk" in config:
     click.echo("Creating additional resources [cdk]...")
     # Import the project's infra module
-    project_infra_path = os.path.join(project_dir, config["cdk"]["path_to_infra"])
-    sys.path.append(project_infra_path)
-    from main import custom_infra  # pylint: disable=import-error
+    project_infra_path = Path(project_dir) / config["cdk"]["path_to_infra"]
+    sys.path.append(str(project_infra_path))
+    from main import custom_infra  # type: ignore
 
     custom_infra_config = custom_infra(parent_stack, region)
     modules["custom_infra"] = custom_infra_config
@@ -72,47 +71,52 @@ if "api_gateway" in config:
 if "api" in config:
     click.echo("Creating stack for [api]...")
 
+    gateway: apigw.IRestApi | None = None
+
     if "api_gateway_id" in config["api"]:
         rest_api_id = config["api"]["api_gateway_id"]
         # TODO: Move this to utils.py
         # Get the root resource ID from the API Gateway
         apigw_client = boto3.client("apigateway", region_name=region)
-        response = apigw_client.get_resources(
-            restApiId=rest_api_id
-        )
-        root_resource_id = [
-            resource["id"]
-            for resource in response["items"]
-            if resource["path"] == "/"
+        response = apigw_client.get_resources(restApiId=rest_api_id)
+        root_resource_id: list[str] = [
+            resource["id"] for resource in response["items"] if resource["path"] == "/"
         ]
         if len(root_resource_id) == 0:
-            raise Exception("Could not find root resource ID for API Gateway")
+            raise click.ClickException("Could not find root resource ID for API Gateway")
+
         click.echo(f"Found API Gateway root resource ID: {root_resource_id[0]}")
-        root_resource_id = root_resource_id[0]
 
         gateway = apigw.RestApi.from_rest_api_attributes(
             parent_stack,
             "RestApiGatewayReference",
             rest_api_id=rest_api_id,
-            root_resource_id=root_resource_id,
+            root_resource_id=root_resource_id[0],
         )
-    else:
-        gateway = modules["api_gateway"].api if "api_gateway" in modules else None
+    elif "api_gateway" in modules:
+        gateway = modules["api_gateway"].api
 
-    
-    user_pool = parent_stack.node.try_find_child("UserPool") if "api_gateway" in modules else (
-        cognito.UserPool.from_user_pool_id(
-            parent_stack, "UserPool", config["api"]["user_pool_id"]
+    user_pool = (
+        parent_stack.node.try_find_child("UserPool")  # type: ignore
+        if "api_gateway" in modules
+        else (
+            cognito.UserPool.from_user_pool_id(
+                parent_stack, "UserPool", config["api"]["user_pool_id"]
+            )
+            if "user_pool_id" in config["api"]
+            else None
         )
-        if "user_pool_id" in config["api"]
-        else None
     )
 
     click.echo(f"API is passed a UserPool?: {user_pool is not None}")
 
     api_construct_name = config["api"]["name"].replace(" ", "-")
     api_construct = RestApiBackend(
-        parent_stack, api_construct_name, config, rest_api_gw=gateway, user_pool=user_pool
+        parent_stack,
+        api_construct_name,
+        config,
+        rest_api_gw=gateway,  # type: ignore
+        user_pool=user_pool,
     )
 
     # Add custom environment variables if provided
