@@ -1,8 +1,12 @@
-import click
+"""REST API Gateway infrastructure stack."""
 
+from pathlib import Path
+from typing import Any
+
+import click
 from aws_cdk import aws_apigateway as apigw
-from aws_cdk import aws_lambda as lambda_
 from aws_cdk import aws_certificatemanager as acm
+from aws_cdk import aws_lambda as lambda_
 from aws_cdk import aws_route53 as route53
 from aws_cdk import aws_route53_targets as targets
 from aws_solutions_constructs import aws_apigateway_lambda as apigw_lambda
@@ -12,41 +16,66 @@ from kegstandcli.utils import hosted_zone_from_domain
 
 MODULE_CONFIG_KEY = "api_gateway"
 
-class RestApiGateway(Construct):
-    def __init__(self, scope: Construct, id: str, config: dict, user_pool) -> None:
-        super().__init__(scope, id)
 
-        provision_with_authorizer = user_pool is not None
+class RestApiGateway(Construct):
+    """Construct for creating a REST API Gateway.
+
+    This construct creates an API Gateway with optional custom domain name and
+    health check endpoint. It can be configured with or without Cognito user pool
+    authorization.
+    """
+
+    def __init__(
+        self,
+        scope: Construct,
+        construct_id: str,
+        config: dict[str, Any],
+        user_pool: Any | None,  # noqa: ARG002
+    ) -> None:
+        """Initialize the REST API Gateway.
+
+        Args:
+            scope: CDK construct scope
+            construct_id: Unique identifier for the construct
+            config: Project configuration dictionary
+            user_pool: Optional Cognito user pool for authorization
+        """
+        super().__init__(scope, construct_id)
+
+        # provision_with_authorizer = user_pool is not None
+
+        api_gw_src_path = Path(config["project_dir"]) / "dist" / "api_gw_src"
 
         # Lambda API backend
-        default_function_props=lambda_.FunctionProps(
-            function_name=f"{id}-DefaultGatewayFunction",
+        default_function_props = lambda_.FunctionProps(
+            function_name=f"{construct_id}-DefaultGatewayFunction",
             runtime=lambda_.Runtime.PYTHON_3_9,
             handler="api.lambda.handler",
-            code=lambda_.Code.from_asset(f'{config["project_dir"]}/dist/api_gw_src'),
+            code=lambda_.Code.from_asset(str(api_gw_src_path)),
             memory_size=256,
             tracing=lambda_.Tracing.ACTIVE,
             environment={
                 "LOG_LEVEL": "INFO",
-                "POWERTOOLS_LOGGER_SAMPLE_RATE": "1.00", # "0.05",  # Use log level DEBUG for 5% of invocations
+                "POWERTOOLS_LOGGER_SAMPLE_RATE": "1.00",
                 "POWERTOOLS_LOGGER_LOG_EVENT": "true",
-                "POWERTOOLS_SERVICE_NAME": f"{id}-DefaultGatewayFunction",
+                "POWERTOOLS_SERVICE_NAME": f"{construct_id}-DefaultGatewayFunction",
             },
         )
 
         health_lambda_function = lambda_.Function(
-            self, f"{id}-HealthCheckFunction",
-            function_name=f"{id}-HealthCheckFunction",
+            self,
+            f"{construct_id}-HealthCheckFunction",
+            function_name=f"{construct_id}-HealthCheckFunction",
             runtime=lambda_.Runtime.PYTHON_3_9,
             handler="api.lambda.handler",
-            code=lambda_.Code.from_asset(f'{config["project_dir"]}/dist/api_gw_src'),
+            code=lambda_.Code.from_asset(str(api_gw_src_path)),
             memory_size=256,
             tracing=lambda_.Tracing.ACTIVE,
             environment={
                 "LOG_LEVEL": "INFO",
-                "POWERTOOLS_LOGGER_SAMPLE_RATE": "1.00", # "0.05",  # Use log level DEBUG for 5% of invocations
+                "POWERTOOLS_LOGGER_SAMPLE_RATE": "1.00",
                 "POWERTOOLS_LOGGER_LOG_EVENT": "true",
-                "POWERTOOLS_SERVICE_NAME": f"{id}-HealthCheckFunction",
+                "POWERTOOLS_SERVICE_NAME": f"{construct_id}-HealthCheckFunction",
             },
         )
 
@@ -55,13 +84,14 @@ class RestApiGateway(Construct):
         if "domain_name" in config[MODULE_CONFIG_KEY]:
             # Return an error if domain_certificate_arn is not specified
             if "domain_certificate_arn" not in config[MODULE_CONFIG_KEY]:
-                click.ClickException(
-                    "Config [api_gateway].domain_certificate_arn must be specified when using a custom domain name."
+                raise click.ClickException(
+                    "Config [api_gateway].domain_certificate_arn must be "
+                    "specified when using a custom domain name."
                 )
 
             # API Gateway w. custom domain name
-            api_gateway_props=apigw.LambdaRestApiProps(
-                rest_api_name=f"{id}-RestApi",
+            api_gateway_props = apigw.LambdaRestApiProps(
+                rest_api_name=f"{construct_id}-RestApi",
                 handler=health_lambda_function,
                 proxy=False,  # Disable default proxy resource
                 default_method_options=apigw.MethodOptions(
@@ -86,14 +116,18 @@ class RestApiGateway(Construct):
             # https://docs.aws.amazon.com/solutions/latest/constructs/aws-apigateway-lambda.html
             api_gateway_to_lambda = apigw_lambda.ApiGatewayToLambda(
                 self,
-                f"{id}-ApiGatewayConstruct",
+                f"{construct_id}-ApiGatewayConstruct",
                 lambda_function_props=default_function_props,
                 api_gateway_props=api_gateway_props,
             )
             api = api_gateway_to_lambda.api_gateway
 
             # Add the Route53 record for the API subdomain
-            hosted_zone = route53.HostedZone.from_lookup(self, "HostedZone", domain_name=hosted_zone_from_domain(config[MODULE_CONFIG_KEY]["domain_name"]))
+            hosted_zone = route53.HostedZone.from_lookup(
+                self,
+                "HostedZone",
+                domain_name=hosted_zone_from_domain(config[MODULE_CONFIG_KEY]["domain_name"]),
+            )
             route53.ARecord(
                 self,
                 "ApiSubdomainDnsRecord",
@@ -104,8 +138,8 @@ class RestApiGateway(Construct):
 
         else:
             # API Gateway w/o custom domain name
-            api_gateway_props=apigw.LambdaRestApiProps(
-                rest_api_name=f"{id}-RestApi",
+            api_gateway_props = apigw.LambdaRestApiProps(
+                rest_api_name=f"{construct_id}-RestApi",
                 handler=health_lambda_function,
                 proxy=False,  # Disable default proxy resource
                 default_method_options=apigw.MethodOptions(
@@ -115,14 +149,14 @@ class RestApiGateway(Construct):
                     logging_level=apigw.MethodLoggingLevel.INFO,
                     metrics_enabled=True,
                     tracing_enabled=True,
-                )
+                ),
             )
 
             # Official 'ApiGatewayToLambda' AWS Solution Construct
             # https://docs.aws.amazon.com/solutions/latest/constructs/aws-apigateway-lambda.html
             api_gateway_to_lambda = apigw_lambda.ApiGatewayToLambda(
                 self,
-                f"{id}-ApiGatewayConstruct",
+                f"{construct_id}-ApiGatewayConstruct",
                 lambda_function_props=default_function_props,
                 api_gateway_props=api_gateway_props,
             )
@@ -134,4 +168,4 @@ class RestApiGateway(Construct):
         resource_root = self.api.root.add_resource("health")
         resource_root.add_method("GET", apigw.LambdaIntegration(health_lambda_function))
 
-        self.deployment = apigw.Deployment(self, f"{id}-Deployment", api=self.api)
+        self.deployment = apigw.Deployment(self, f"{construct_id}-Deployment", api=self.api)
